Disqus Integration Attempt
Disqus has been integrated into the React app using a solution found here: https://blog.m4x.io/add-disqus-to-your-react-app-in-a-glance/ A small modification was added for 2 inline styles to set the width of the Disqus element. Currently the old comments are not appearing, however I believe this is due to the app running on localhost.