Snack bar notifications
This commit includes all the code for the new snack bar notification system. Using a hidden tag in the Ghost admin we are now able to have a snack bar appear on the site. The intended use is to alert users when we've launched a new update and direct them to the changelog